+++ src/main/java/com/example/itspower/service/GroupRoleService.java
@@ -1,14 +1,11 @@
 package com.example.itspower.service;
-
 import com.example.itspower.model.entity.GroupEntity;
 import com.example.itspower.request.GroupRoleRequest;
 import com.example.itspower.response.dynamic.PageResponse;
 import com.example.itspower.response.group.GroupRoleDemarcationRes;
 import com.example.itspower.response.group.GroupRoleResponse;
 import com.example.itspower.response.group.ResponseCount;
-
 import java.util.List;
-
 public interface GroupRoleService {
     List<GroupRoleResponse> searchAll();
 
@@ -27,7 +24,7 @@ public interface GroupRoleService {
     Object getDemarcationRes(Integer groupId);
     List<GroupRoleDemarcationRes> getAllDemarcationRes();
 
-    Object updateGroupRole(Integer id, Integer demarcation);
+    Object updateGroupRole(Integer id, Float demarcation);
 
     void delete(Integer groupId);
